Increase cookie range
For higher security I changed the cookie type  to long and doing a
randomSeed(micros()) every time before generating the cookie. I am also
sending the cookie in hexadecimal format so that it is only 8 digits max
instead of 10. This seems to have slowed down the authentication process
significantly.

+++ EtherEvent.h
@@ -20,7 +20,7 @@ class EtherEventClass {
   public:
     EtherEventClass();
     boolean begin(const byte eventLengthMaxInput = 15, const byte payloadLengthMaxInput = 100);  //these are the default max length values
-    byte availableEvent(EthernetServer &ethernetServer, int cookieInput = false);
+    byte availableEvent(EthernetServer &ethernetServer, long cookieInput = false);
     byte availablePayload();
     void readEvent(char eventBuffer[]);
     void readPayload(char payloadBuffer[]);
